trade/requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0

# Market data APIs
yfinance>=0.2.40
alpha-vantage>=2.3.1

# Technical analysis
ta>=0.11.0
pandas-ta>=0.3.14b0

# Sentiment analysis
textblob>=0.17.1
newsapi-python>=0.2.7

# Data visualization (optional)
matplotlib>=3.8.0

# Configuration
python-dotenv>=1.0.0

# Timezone handling
pytz>=2024.1

# Web interface
flask>=3.0.0

# Advanced calculations
scipy>=1.11.0
